% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdnada-package.r
\docType{package}
\name{sdnada}
\alias{sdnada}
\alias{sdnada-package}
\title{sdnada.}
\description{
Contains contrived functions that compute the standard deviation or standard error of a numeric vector.
}

