% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_deviation.R
\name{standard_deviation}
\alias{standard_deviation}
\title{Standard Deviation}
\usage{
standard_deviation(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
numeric
}
\description{
This contrived function computes the standard deviation of the values in x.
}
\examples{
standard_deviation(c(1, 7, 3.7, 2, 0.3))
}

