% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_deviation.R
\name{standard_error}
\alias{standard_error}
\title{Standard Error}
\usage{
standard_error(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
numeric
}
\description{
This contrived function computes the standard error of the values in x.
}
\examples{
standard_error(c(1, 7, 3.7, 2, 0.3))
}

